fase 2: split ViewModel
