+++ AutoDarkModeApp/MainWindow.xaml
@@ -68,7 +68,7 @@
                 <NavigationViewItem
                     Content="{helpers:ResourceString Name=AutoSwitch}"
                     Icon="{ui:FontIcon Glyph=&#xE823;}"
-                    Tag="Time" />
+                    Tag="AutoSwitch" />
                 <NavigationViewItem
                     Content="{helpers:ResourceString Name=SystemAreas}"
                     Icon="{ui:FontIcon Glyph=&#xEC4E;}"
