+++ AutoDarkModeApp/Views/AutoSwitchPage.xaml.cs
@@ -7,11 +7,11 @@ namespace AutoDarkModeApp.Views;
 
 public sealed partial class AutoSwitchPage : Page
 {
-    public TimeViewModel ViewModel { get; }
+    public AutoSwitchViewModel ViewModel { get; }
 
     public AutoSwitchPage()
     {
-        ViewModel = App.GetService<TimeViewModel>();
+        ViewModel = App.GetService<AutoSwitchViewModel>();
         InitializeComponent();
         ViewModel.PropertyChanged += ViewModel_PropertyChanged;
         Unloaded += (s, e) => ViewModel.PropertyChanged -= ViewModel_PropertyChanged;
